+++ database/migrations/2017_03_19_235409_add_fields_to_reservation_options_table.php
@@ -14,8 +14,12 @@ class AddFieldsToReservationOptionsTable extends Migration
     {
         Schema::table('reservations_options', function (Blueprint $table) {
             //
-            $table->decimal('price_from', 5, 2)->unsigned()->nullable();
-            $table->decimal('price', 5, 2)->unsigned()->nullable();
+            if(!Schema::hasColumn('reservations_options', 'price_from')){
+                $table->decimal('price_from', 5, 2)->unsigned()->nullable();
+            }
+            if(!Schema::hasColumn('reservations_options', 'price')){
+                $table->decimal('price', 5, 2)->unsigned()->nullable();
+            }            
         });
     }
 
@@ -28,8 +32,12 @@ class AddFieldsToReservationOptionsTable extends Migration
     {
         Schema::table('reservations_options', function (Blueprint $table) {
             //
-            $table->dropColumn('price_from');
-            $table->dropColumn('price');
+            if(Schema::hasColumn('reservations_options', 'price_from')){
+                $table->dropColumn('price_from');
+            }
+            if(Schema::hasColumn('reservations_options', 'price')){
+                $table->dropColumn('price');
+            }            
         });
     }
 }
